main: use int for representing cork queue indexes
Close #2490.

unsigned int was used for them in some areas of code mistakenly.
This change unifies the type for the indexes to int.

TODO: handling the case that makeTagEntry returns CORK_NIL.  If more
than INT_MAX tag entries are stored to the cork queue, makeTagEntry
returns CORK_NIL. Many parsers assumes makeTagEntry never returns
CORK_NIL. Ths assumption can be cause a crash of ctags process.

+++ parsers/python.c
@@ -1227,7 +1227,7 @@ static void setIndent (tokenInfo *const token)
 	{
 		if (lv->corkIndex != CORK_NIL)
 		{
-			tagEntryInfo *e = getEntryInCorkQueue ((unsigned int) lv->corkIndex);
+			tagEntryInfo *e = getEntryInCorkQueue (lv->corkIndex);
 
 			e->extensionFields.endLine = token->lineNumber;
 		}
